Change terrain indices to be 2 bytes and fix mouse picker

+++ barbet/src/main/game-state/world/world-to-mesh-converter.ts
@@ -247,7 +247,7 @@ export const buildChunkMesh = (world: WorldLike, chunkX: number, chunkZ: number,
 
 	return {
 		vertexes: finalVertexes,
-		indices: new Uint32Array(indices),
+		indices: new Uint16Array(indices),
 	}
 }
